orientation-event/DeviceOrientation/t009.html

<!DOCTYPE html>
<html>

  <head>
    <title>Rotate the device frame around its x axis:</title>
	<meta name=viewport content="width=device-width, maximum-scale=1.0, user-scalable=no" />
    <script src="http://w3c-test.org/resources/testharness.js"></script>
    <script src="http://w3c-test.org/resources/testharnessreport.js"></script>
    <link rel='author' title='Mosquito FP7'/>
  </head>

  <body>

  
  		<h1>Rotate the device frame around its x axis:</h1>
    <div id="log">
    </div>
	
	<script>
  		window.addEventListener('deviceorientation', function(eventData) {
			beta = eventData.beta;
		}, false);
	
		alert("Launch Test.");
		test(function() {assert_true(beta > -180 && beta < 180)},"Beta is between -180 and 180");
		test(function() {assert_false(beta > -2 && beta < 2)},"Beta is not around zero (between -2 and 2)");
		  
	</script>
	
  </body>
  
</html>
